Modified : table fields and get user id in token form

+++ src/main/java/com/bridgelabz/demo/service/LabelService.java
@@ -1,5 +1,6 @@
 package com.bridgelabz.demo.service;
 
+import java.time.LocalDateTime;
 import java.util.List;
 import java.util.NoSuchElementException;
 
@@ -12,12 +13,12 @@ import org.springframework.stereotype.Service;
 import com.bridgelabz.demo.dto.Response;
 import com.bridgelabz.demo.enumeration.Message;
 import com.bridgelabz.demo.model.Label;
-import com.bridgelabz.demo.model.LabelNotesMapping;
 import com.bridgelabz.demo.model.Note;
-import com.bridgelabz.demo.repository.LabelNotesMappingRepository;
+import com.bridgelabz.demo.model.User;
 import com.bridgelabz.demo.repository.LabelRepository;
 import com.bridgelabz.demo.repository.NoteRepository;
 import com.bridgelabz.demo.repository.UserRepository;
+import com.bridgelabz.demo.util.UserToken;
 
 @Service
 public class LabelService {
@@ -29,9 +30,6 @@ public class LabelService {
 	private LabelRepository labelRepository;
 
 	@Autowired
-	private LabelNotesMappingRepository labelNotesMappingRepository;
-
-	@Autowired
 	private UserRepository userRepository;
 
 	@Autowired
@@ -41,10 +39,12 @@ public class LabelService {
 
 	}
 
-	public ResponseEntity<Response> addLabel(Label label) {
+	public ResponseEntity<Response> addLabel(Label label, String token) {
+		Long userId = UserToken.getUserIdFromToken(token);
 		try {
-			userRepository.findById(label.getUserId()).get();
-			List<Label> allLabels = labelRepository.findAllByUserId(label.getUserId());
+			User user = userRepository.findById(userId).get();
+			label.setUser(user);
+			List<Label> allLabels = labelRepository.findAllLabelsByUserId(userId);
 			if (allLabels.contains(label))
 				return new ResponseEntity<Response>(userService.getResponse(Message.LABEL_NAME_EXISTS, null, 409),
 						HttpStatus.CONFLICT);
@@ -61,23 +61,19 @@ public class LabelService {
 				HttpStatus.SERVICE_UNAVAILABLE);
 	}
 
-	public ResponseEntity<Response> updateLabel(Label label) {
+	public ResponseEntity<Response> updateLabel(Label label, String token) {
+		Long userId = UserToken.getUserIdFromToken(token);
 		try {
-			userRepository.findById(label.getUserId()).get();
-		} catch (NoSuchElementException e) {
-			return new ResponseEntity<Response>(userService.getResponse(Message.USER_ID_DOESNOT_EXISTS, null, 404),
-					HttpStatus.NOT_FOUND);
-		}
-		try {
-			if (labelRepository.findById(label.getId()).get() != null) {
-				List<Label> allLabels = labelRepository.findAllByUserId(label.getUserId());
-				if (allLabels.contains(label))
-					return new ResponseEntity<Response>(userService.getResponse(Message.LABEL_NAME_EXISTS, null, 409),
-							HttpStatus.CONFLICT);
-				Label labels = labelRepository.save(label);
+			Label labels = labelRepository.findById(label.getLabelId()).get();
+			if (labels.getUser().getUserId() == userId) {
+				labels.setModifiedDate(LocalDateTime.now());
+				labels.setLabelName(label.getLabelName());
+				labels = labelRepository.save(labels);
 				return new ResponseEntity<Response>(userService.getResponse(Message.SUCCESSFUL, labels, 200),
 						HttpStatus.OK);
 			}
+			return new ResponseEntity<Response>(userService.getResponse(Message.INVALID_USER_ID, null, 404),
+					HttpStatus.BAD_REQUEST);
 		} catch (NoSuchElementException e) {
 			return new ResponseEntity<Response>(userService.getResponse(Message.LABEL_ID_DOESNOT_EXISTS, null, 404),
 					HttpStatus.NOT_FOUND);
@@ -88,13 +84,18 @@ public class LabelService {
 				HttpStatus.SERVICE_UNAVAILABLE);
 	}
 
-	public ResponseEntity<Response> deleteLabel(Long labelId) {
+	public ResponseEntity<Response> deleteLabel(Long labelId, String token) {
+		Long userId = UserToken.getUserIdFromToken(token);
 		try {
-			if (labelNotesMappingRepository.findByLabelId(labelId) != null)
-				labelNotesMappingRepository.deleteByLabelId(labelId);
-			labelRepository.deleteById(labelId);
-			return new ResponseEntity<Response>(
-					userService.getResponse(Message.SUCCESSFUL, Message.LABEL_DELETED_SUCCESFULLY, 200), HttpStatus.OK);
+			Label label = labelRepository.findById(labelId).get();
+			if (label.getUser().getUserId().equals(userId)) {
+				labelRepository.deleteById(labelId);
+				return new ResponseEntity<Response>(
+						userService.getResponse(Message.SUCCESSFUL, Message.LABEL_DELETED_SUCCESFULLY, 200),
+						HttpStatus.OK);
+			}
+			return new ResponseEntity<Response>(userService.getResponse(Message.INVALID_USER_ID, null, 404),
+					HttpStatus.BAD_REQUEST);
 		} catch (EmptyResultDataAccessException e) {
 			return new ResponseEntity<Response>(userService.getResponse(Message.LABEL_ID_DOESNOT_EXISTS, null, 404),
 					HttpStatus.NOT_FOUND);
@@ -105,10 +106,11 @@ public class LabelService {
 				HttpStatus.SERVICE_UNAVAILABLE);
 	}
 
-	public ResponseEntity<Response> getAllLabelsByUserId(Long userId) {
+	public ResponseEntity<Response> getAllLabelsByUserId(String token) {
+		Long userId = UserToken.getUserIdFromToken(token);
 		try {
 			userRepository.findById(userId).get();
-			List<Label> allLabels = labelRepository.findAllByUserId(userId);
+			List<Label> allLabels = labelRepository.findAllLabelsByUserId(userId);
 			return new ResponseEntity<Response>(userService.getResponse(Message.SUCCESSFUL, allLabels, 200),
 					HttpStatus.OK);
 		} catch (NoSuchElementException e) {
@@ -121,22 +123,24 @@ public class LabelService {
 				HttpStatus.SERVICE_UNAVAILABLE);
 	}
 
-	public ResponseEntity<Response> mapLabelAndNotes(LabelNotesMapping labelNotesMapping) {
+	public ResponseEntity<Response> mapLabelAndNotes(String token, Long labelId, Long noteId) {
 		Label label;
 		try {
-			label = labelRepository.findById(labelNotesMapping.getLabelId()).get();
+			label = labelRepository.findById(labelId).get();
 		} catch (NoSuchElementException e) {
 			return new ResponseEntity<Response>(userService.getResponse(Message.LABEL_ID_DOESNOT_EXISTS, null, 404),
 					HttpStatus.NOT_FOUND);
 		}
 		try {
-			Note note = noteRepository.findById(labelNotesMapping.getNoteId()).get();
-			if (!note.getUserId().equals(label.getUserId()))
+			Note note = noteRepository.findById(noteId).get();
+			if (!note.getUser().getUserId().equals(label.getUser().getUserId()))
 				return new ResponseEntity<Response>(
-						userService.getResponse(Message.USER_ID_OF_LABEL_DOESNT_MATCH_WITH_NOTE_USER_ID, null, 200),
-						HttpStatus.OK);
-			LabelNotesMapping labelNotesMappings = labelNotesMappingRepository.save(labelNotesMapping);
-			return new ResponseEntity<Response>(userService.getResponse(Message.SUCCESSFUL, labelNotesMappings, 200),
+						userService.getResponse(Message.USER_ID_OF_LABEL_DOESNT_MATCH_WITH_NOTE_USER_ID, null, 404),
+						HttpStatus.BAD_REQUEST);
+			note.getLabel().add(label);
+			note.setModifiedDate(LocalDateTime.now());
+			note = noteRepository.save(note);
+			return new ResponseEntity<Response>(userService.getResponse(Message.SUCCESSFUL, note, 200),
 					HttpStatus.OK);
 		} catch (NoSuchElementException e) {
 			return new ResponseEntity<Response>(userService.getResponse(Message.NOTE_ID_DOESNOT_EXISTS, null, 404),
