+++ src/main/java/com/bridgelabz/demo/service/NoteService.java
@@ -1,7 +1,10 @@
 package com.bridgelabz.demo.service;
 
+import java.time.LocalDateTime;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.NoSuchElementException;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -10,26 +13,20 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
+import com.bridgelabz.demo.dto.Collabrator;
 import com.bridgelabz.demo.dto.Response;
 import com.bridgelabz.demo.enumeration.Message;
-import com.bridgelabz.demo.model.Collabrator;
 import com.bridgelabz.demo.model.Note;
-import com.bridgelabz.demo.repository.CollabratorRepository;
-import com.bridgelabz.demo.repository.LabelNotesMappingRepository;
+import com.bridgelabz.demo.model.User;
 import com.bridgelabz.demo.repository.NoteRepository;
 import com.bridgelabz.demo.repository.UserRepository;
+import com.bridgelabz.demo.util.UserToken;
 
 @Service
 public class NoteService {
 
 	@Autowired
 	private NoteRepository noteRepository;
-
-	@Autowired
-	private LabelNotesMappingRepository labelNotesMappingRepository;
-
-	@Autowired
-	private CollabratorRepository collabratorRepository;
 
 	@Autowired
 	private UserRepository userRepository;
@@ -41,9 +38,11 @@ public class NoteService {
 
 	}
 
-	public ResponseEntity<Response> addNote(Note note) {
+	public ResponseEntity<Response> addNote(Note note, String token) {
+		Long userId = UserToken.getUserIdFromToken(token);
 		try {
-			userRepository.findById(note.getUserId()).get();
+			User user = userRepository.findById(userId).get();
+			note.setUser(user);
 			Note noteDetails = noteRepository.save(note);
 			return new ResponseEntity<Response>(
 					userService.getResponse(Message.NOTE_ADDED_SUCCESFULLY, noteDetails, 200), HttpStatus.OK);
@@ -57,11 +56,15 @@ public class NoteService {
 				HttpStatus.SERVICE_UNAVAILABLE);
 	}
 
-	public ResponseEntity<Response> getAllNotesByUserId(Long userId, boolean isTrash) {
+	public ResponseEntity<Response> getAllNotesByUserId(String token, boolean isTrash) {
+		List<Note> availableNotes = new ArrayList<Note>();
+		Long userId = UserToken.getUserIdFromToken(token);
 		try {
-			userRepository.findById(userId).get();
-			List<Note> availableNotes = noteRepository.findAllByUserId(userId).stream()
-					.filter(note -> note.getTrash() == isTrash && !note.getArchive()).collect(Collectors.toList());
+			Optional<User> user = userRepository.findById(userId);
+			List<Note> collaboratedNote = user.get().getCollaboratedNote().stream().collect(Collectors.toList());
+			availableNotes = noteRepository.findAllNotesByUserId(userId, Boolean.valueOf(false),
+					Boolean.valueOf(isTrash));
+			availableNotes.addAll(collaboratedNote);
 			return new ResponseEntity<Response>(userService.getResponse(Message.SUCCESSFUL, availableNotes, 200),
 					HttpStatus.OK);
 		} catch (NoSuchElementException e) {
@@ -74,11 +77,10 @@ public class NoteService {
 				HttpStatus.SERVICE_UNAVAILABLE);
 	}
 
-	public ResponseEntity<Response> getAllArchiveNotesByUserId(Long userId) {
+	public ResponseEntity<Response> getAllArchiveNotesByUserId(String token) {
+		Long userId = UserToken.getUserIdFromToken(token);
 		try {
-			userRepository.findById(userId).get();
-			List<Note> availableNotes = noteRepository.findAllByUserId(userId).stream()
-					.filter(note -> !note.getTrash() && note.getArchive()).collect(Collectors.toList());
+			List<Note> availableNotes = noteRepository.findAllNotesByUserId(userId, true, false);
 			return new ResponseEntity<Response>(userService.getResponse(Message.SUCCESSFUL, availableNotes, 200),
 					HttpStatus.OK);
 		} catch (NoSuchElementException e) {
@@ -91,12 +93,18 @@ public class NoteService {
 				HttpStatus.SERVICE_UNAVAILABLE);
 	}
 
-	public ResponseEntity<Response> setIsTrash(Long noteId, boolean isTrash) {
+	public ResponseEntity<Response> setIsTrash(Long noteId, boolean isTrash, String token) {
+		Long userId = UserToken.getUserIdFromToken(token);
 		try {
 			Note note = noteRepository.findById(noteId).get();
-			note.setTrash(isTrash);
-			Note notes = noteRepository.save(note);
-			return new ResponseEntity<Response>(userService.getResponse(Message.SUCCESSFUL, notes, 200), HttpStatus.OK);
+			if (note.getUser().getUserId().equals(userId)) {
+				note.setTrash(isTrash);
+				Note notes = noteRepository.save(note);
+				return new ResponseEntity<Response>(userService.getResponse(Message.SUCCESSFUL, notes, 200),
+						HttpStatus.OK);
+			}
+			return new ResponseEntity<Response>(userService.getResponse(Message.INVALID_USER_ID, null, 404),
+					HttpStatus.BAD_REQUEST);
 		} catch (NoSuchElementException e) {
 			return new ResponseEntity<Response>(userService.getResponse(Message.NOTE_ID_DOESNOT_EXISTS, null, 404),
 					HttpStatus.NOT_FOUND);
@@ -107,15 +115,18 @@ public class NoteService {
 				HttpStatus.SERVICE_UNAVAILABLE);
 	}
 
-	public ResponseEntity<Response> deleteNote(Long noteId) {
+	public ResponseEntity<Response> deleteNote(Long noteId, String token) {
+		Long userId = UserToken.getUserIdFromToken(token);
 		try {
-			if (labelNotesMappingRepository.findByNoteId(noteId) != null)
-				labelNotesMappingRepository.deleteByNoteId(noteId);
-			if (collabratorRepository.findByNoteId(noteId) != null)
-				collabratorRepository.deleteByNoteId(noteId);
-			noteRepository.deleteById(noteId);
-			return new ResponseEntity<Response>(
-					userService.getResponse(Message.SUCCESSFUL, Message.NOTE_DELETED_SUCCESFULLY, 200), HttpStatus.OK);
+			Note note = noteRepository.findById(noteId).get();
+			if (note.getUser().getUserId().equals(userId)) {
+				noteRepository.deleteById(noteId);
+				return new ResponseEntity<Response>(
+						userService.getResponse(Message.SUCCESSFUL, Message.NOTE_DELETED_SUCCESFULLY, 200),
+						HttpStatus.OK);
+			}
+			return new ResponseEntity<Response>(userService.getResponse(Message.INVALID_USER_ID, null, 404),
+					HttpStatus.BAD_REQUEST);
 		} catch (EmptyResultDataAccessException e) {
 			return new ResponseEntity<Response>(userService.getResponse(Message.NOTE_ID_DOESNOT_EXISTS, null, 404),
 					HttpStatus.NOT_FOUND);
@@ -126,18 +137,20 @@ public class NoteService {
 				HttpStatus.SERVICE_UNAVAILABLE);
 	}
 
-	public ResponseEntity<Response> updateNote(Note note) {
+	public ResponseEntity<Response> updateNote(Note note, String token) {
+		Long userId = UserToken.getUserIdFromToken(token);
 		try {
-			userRepository.findById(note.getUserId()).get();
-		} catch (NoSuchElementException e) {
-			return new ResponseEntity<Response>(userService.getResponse(Message.USER_ID_DOESNOT_EXISTS, null, 404),
-					HttpStatus.NOT_FOUND);
-		}
-		try {
-			noteRepository.findById(note.getId()).get();
-			Note noteDetails = noteRepository.save(note);
-			return new ResponseEntity<Response>(
-					userService.getResponse(Message.NOTE_UPDATED_SUCCESFULLY, noteDetails, 200), HttpStatus.OK);
+			Note notes = noteRepository.findById(note.getNoteId()).get();
+			if (notes.getUser().getUserId().equals(userId)) {
+				notes.setTitle(note.getTitle());
+				notes.setDescription(note.getDescription());
+				notes.setModifiedDate(LocalDateTime.now());
+				Note noteDetails = noteRepository.save(notes);
+				return new ResponseEntity<Response>(
+						userService.getResponse(Message.NOTE_UPDATED_SUCCESFULLY, noteDetails, 200), HttpStatus.OK);
+			}
+			return new ResponseEntity<Response>(userService.getResponse(Message.INVALID_USER_ID, null, 404),
+					HttpStatus.BAD_REQUEST);
 		} catch (NoSuchElementException e) {
 			return new ResponseEntity<Response>(userService.getResponse(Message.NOTE_ID_DOESNOT_EXISTS, null, 404),
 					HttpStatus.NOT_FOUND);
@@ -148,12 +161,22 @@ public class NoteService {
 				HttpStatus.SERVICE_UNAVAILABLE);
 	}
 
-	public ResponseEntity<Response> addCollabrator(Collabrator collabrator) {
+	public ResponseEntity<Response> addCollabrator(Collabrator collabrator, String token) {
+		Long userId = UserToken.getUserIdFromToken(token);
 		try {
-			noteRepository.findById(collabrator.getNoteId()).get();
-			Collabrator collabrators = collabratorRepository.save(collabrator);
-			return new ResponseEntity<Response>(userService.getResponse(Message.SUCCESSFUL, collabrators, 200),
-					HttpStatus.OK);
+			User user = userRepository.findByEmail(collabrator.getCollabratorEmail());
+			Note note = noteRepository.findById(collabrator.getNoteId()).get();
+			if (userId == note.getUser().getUserId()) {
+				user.getCollaboratedNote().add(note);
+				note.getCollaboratedUser().add(user);
+				note.setModifiedDate(LocalDateTime.now());
+				userRepository.save(user);
+				note = noteRepository.save(note);
+				return new ResponseEntity<Response>(userService.getResponse(Message.SUCCESSFUL, note, 200),
+						HttpStatus.OK);
+			}
+			return new ResponseEntity<Response>(userService.getResponse(Message.INVALID_USER_ID, null, 404),
+					HttpStatus.BAD_REQUEST);
 		} catch (NoSuchElementException e) {
 			return new ResponseEntity<Response>(userService.getResponse(Message.NOTE_ID_DOESNOT_EXISTS, null, 404),
 					HttpStatus.NOT_FOUND);
